completed auth gatekeeping for cart information, also added security against injection attacks for signup page

+++ server/api/cart.js
@@ -47,32 +47,38 @@ router.put('/:userId', requireToken, async (req, res, next) => {
 });
 
 //delete product where user has it
-router.delete('/:userId/:productId', async (req, res, next) => {
+router.delete('/:userId/:productId', requireToken, async (req, res, next) => {
   try {
-    const userId = req.params.userId;
-    const productId = req.params.productId;
-    await Cart.destroy({
-      where: {
-        userId: userId,
-        productId: productId,
-      },
-    });
-    res.sendStatus(204);
+    const paramsNum = parseInt(req.params.userId, 10);
+    if (req.user.id === paramsNum) {
+      const userId = req.params.userId;
+      const productId = req.params.productId;
+      await Cart.destroy({
+        where: {
+          userId: userId,
+          productId: productId,
+        },
+      });
+      res.sendStatus(204);
+    }
   } catch (error) {
     next(error);
   }
 });
 
 //delete the items in cart when user 'checks out' (tier 1 requirements min-maxing)
-router.delete('/:userId/', async (req, res, next) => {
+router.delete('/:userId/', requireToken, async (req, res, next) => {
   try {
-    const userId = req.params.userId;
-    await Cart.destroy({
-      where: {
-        userId: userId,
-      },
-    });
-    res.sendStatus(204);
+    const paramsNum = parseInt(req.params.userId, 10);
+    if (req.user.id === paramsNum) {
+      const userId = req.params.userId;
+      await Cart.destroy({
+        where: {
+          userId: userId,
+        },
+      });
+      res.sendStatus(204);
+    }
   } catch (error) {
     next(error);
   }
@@ -84,43 +90,46 @@ router.delete('/:userId/', async (req, res, next) => {
 //then, add the requested item to the cart
 //then, assign that cart to the requested user
 
-router.post('/:userId', async (req, res, next) => {
+router.post('/:userId', requireToken, async (req, res, next) => {
   try {
-    const productId = req.body.productId;
-    const quantity = req.body.quantity || 1;
-    const userId = req.params.userId;
+    const paramsNum = parseInt(req.params.userId, 10);
+    if (req.user.id === paramsNum) {
+      const productId = req.body.productId;
+      const quantity = req.body.quantity || 1;
+      const userId = req.params.userId;
 
-    let product = await Cart.findOne({
-      where: {
-        userId: userId,
-        productId: productId,
-      },
-    });
+      let product = await Cart.findOne({
+        where: {
+          userId: userId,
+          productId: productId,
+        },
+      });
 
-    if (product) {
-      //this exist already in user cart
-      product.quantity = product.quantity + quantity;
+      if (product) {
+        //this exist already in user cart
+        product.quantity = product.quantity + quantity;
 
-      await product.save();
-      res.sendStatus(201);
-    } else {
-      //this does not exist in user cart
-      const newCartObj = {
-        quantity: quantity,
-      };
-      //create new Cart from model w/ quantity specified
-      const newCart = await Cart.create(newCartObj);
+        await product.save();
+        res.sendStatus(201);
+      } else {
+        //this does not exist in user cart
+        const newCartObj = {
+          quantity: quantity,
+        };
+        //create new Cart from model w/ quantity specified
+        const newCart = await Cart.create(newCartObj);
 
-      //add association to specified user
-      const user = await User.findByPk(userId);
-      await newCart.setUser(user);
+        //add association to specified user
+        const user = await User.findByPk(userId);
+        await newCart.setUser(user);
 
-      //add association to specified product
-      const product = await Product.findByPk(productId);
-      await newCart.setProduct(product);
+        //add association to specified product
+        const product = await Product.findByPk(productId);
+        await newCart.setProduct(product);
 
-      //what should we return? what would JPFP do?
-      res.sendStatus(201);
+        //what should we return? what would JPFP do?
+        res.sendStatus(201);
+      }
     }
   } catch (error) {
     next(error);
